Change style_box to stylebox in file names
Change mentions of style_box in file names to stylebox to maintain consistancy
