+++ editor/plugins/style_box_editor_plugin.h
@@ -28,8 +28,8 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
-#ifndef STYLE_BOX_EDITOR_PLUGIN_H
-#define STYLE_BOX_EDITOR_PLUGIN_H
+#ifndef STYLEBOX_EDITOR_PLUGIN_H
+#define STYLEBOX_EDITOR_PLUGIN_H
 
 #include "editor/editor_inspector.h"
 #include "editor/plugins/editor_plugin.h"
@@ -75,4 +75,4 @@ public:
 	StyleBoxEditorPlugin();
 };
 
-#endif // STYLE_BOX_EDITOR_PLUGIN_H
+#endif // STYLEBOX_EDITOR_PLUGIN_H
