+++ scene/resources/stylebox_line.cpp
