+++ scene/resources/stylebox_texture.h
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/*  stylebox_flat.h                                                      */
+/*  stylebox_texture.h                                                   */
 /**************************************************************************/
 /*                         This file is part of:                          */
 /*                             GODOT ENGINE                               */
@@ -28,91 +28,72 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
-#ifndef STYLE_BOX_FLAT_H
-#define STYLE_BOX_FLAT_H
+#ifndef STYLEBOX_TEXTURE_H
+#define STYLEBOX_TEXTURE_H
 
 #include "scene/resources/stylebox.h"
+#include "scene/resources/texture.h"
 
-class StyleBoxFlat : public StyleBox {
-	GDCLASS(StyleBoxFlat, StyleBox);
+class StyleBoxTexture : public StyleBox {
+	GDCLASS(StyleBoxTexture, StyleBox);
 
-	Color bg_color = Color(0.6, 0.6, 0.6);
-	Color shadow_color = Color(0, 0, 0, 0.6);
-	Color border_color = Color(0.8, 0.8, 0.8);
+public:
+	enum AxisStretchMode {
+		AXIS_STRETCH_MODE_STRETCH,
+		AXIS_STRETCH_MODE_TILE,
+		AXIS_STRETCH_MODE_TILE_FIT,
+	};
 
-	real_t border_width[4] = {};
-	real_t expand_margin[4] = {};
-	real_t corner_radius[4] = {};
-
+private:
+	float expand_margin[4] = {};
+	float texture_margin[4] = {};
+	Rect2 region_rect;
+	Ref<Texture2D> texture;
 	bool draw_center = true;
-	bool blend_border = false;
-	Vector2 skew;
-	bool anti_aliased = true;
-
-	int corner_detail = 8;
-	int shadow_size = 0;
-	Point2 shadow_offset;
-	real_t aa_size = 1;
+	Color modulate = Color(1, 1, 1, 1);
+	AxisStretchMode axis_h = AXIS_STRETCH_MODE_STRETCH;
+	AxisStretchMode axis_v = AXIS_STRETCH_MODE_STRETCH;
 
 protected:
 	virtual float get_style_margin(Side p_side) const override;
 	static void _bind_methods();
-	void _validate_property(PropertyInfo &p_property) const;
 
 public:
-	void set_bg_color(const Color &p_color);
-	Color get_bg_color() const;
+	void set_texture(Ref<Texture2D> p_texture);
+	Ref<Texture2D> get_texture() const;
 
-	void set_border_color(const Color &p_color);
-	Color get_border_color() const;
-
-	void set_border_width_all(int p_size);
-	int get_border_width_min() const;
-
-	void set_border_width(Side p_side, int p_width);
-	int get_border_width(Side p_side) const;
-
-	void set_border_blend(bool p_blend);
-	bool get_border_blend() const;
-
-	void set_corner_radius_all(int radius);
-	void set_corner_radius_individual(const int radius_top_left, const int radius_top_right, const int radius_bottom_right, const int radius_bottom_left);
-	void set_corner_radius(Corner p_corner, const int radius);
-	int get_corner_radius(Corner p_corner) const;
-
-	void set_corner_detail(const int &p_corner_detail);
-	int get_corner_detail() const;
+	void set_texture_margin(Side p_side, float p_size);
+	void set_texture_margin_all(float p_size);
+	void set_texture_margin_individual(float p_left, float p_top, float p_right, float p_bottom);
+	float get_texture_margin(Side p_side) const;
 
 	void set_expand_margin(Side p_expand_side, float p_size);
 	void set_expand_margin_all(float p_expand_margin_size);
 	void set_expand_margin_individual(float p_left, float p_top, float p_right, float p_bottom);
 	float get_expand_margin(Side p_expand_side) const;
 
+	void set_region_rect(const Rect2 &p_region_rect);
+	Rect2 get_region_rect() const;
+
 	void set_draw_center(bool p_enabled);
 	bool is_draw_center_enabled() const;
 
-	void set_skew(Vector2 p_skew);
-	Vector2 get_skew() const;
+	void set_h_axis_stretch_mode(AxisStretchMode p_mode);
+	AxisStretchMode get_h_axis_stretch_mode() const;
 
-	void set_shadow_color(const Color &p_color);
-	Color get_shadow_color() const;
+	void set_v_axis_stretch_mode(AxisStretchMode p_mode);
+	AxisStretchMode get_v_axis_stretch_mode() const;
 
-	void set_shadow_size(const int &p_size);
-	int get_shadow_size() const;
-
-	void set_shadow_offset(const Point2 &p_offset);
-	Point2 get_shadow_offset() const;
-
-	void set_anti_aliased(const bool &p_anti_aliased);
-	bool is_anti_aliased() const;
-	void set_aa_size(const real_t p_aa_size);
-	real_t get_aa_size() const;
+	void set_modulate(const Color &p_modulate);
+	Color get_modulate() const;
 
 	virtual Rect2 get_draw_rect(const Rect2 &p_rect) const override;
 	virtual void draw(RID p_canvas_item, const Rect2 &p_rect) const override;
 
-	StyleBoxFlat();
-	~StyleBoxFlat();
+	StyleBoxTexture();
+	~StyleBoxTexture();
 };
 
-#endif // STYLE_BOX_FLAT_H
+VARIANT_ENUM_CAST(StyleBoxTexture::AxisStretchMode)
+
+#endif // STYLEBOX_TEXTURE_H
