+++ tests/scene/test_style_box_texture.h
@@ -28,8 +28,8 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
-#ifndef TEST_STYLE_BOX_TEXTURE_H
-#define TEST_STYLE_BOX_TEXTURE_H
+#ifndef TEST_STYLEBOX_TEXTURE_H
+#define TEST_STYLEBOX_TEXTURE_H
 
 #include "scene/resources/stylebox_texture.h"
 
@@ -191,4 +191,4 @@ TEST_CASE("[StyleBoxTexture] get_draw_rect") {
 
 } // namespace TestStyleBoxTexture
 
-#endif // TEST_STYLE_BOX_TEXTURE_H
+#endif // TEST_STYLEBOX_TEXTURE_H
